default to assuming ident auth rather than superuser
Prior to this commit, if no user was specified, the 'postgres' superuser was set as a default. A saner default would be: if no user is specified, we assume the user is using postgres' ident authentication, and does not actually want to explicitly specify a user, instead using the operating system's user.